refactor: re package structure
