+++ swagger-viewer/app/scripts/background.ts
@@ -1,5 +1,5 @@
-import { EXEC_CONVERT_SWAGGER } from "./constants/SendMessageTypes"
-import { ExecConvertSwaggerMessage } from "./types/SendMessage"
+import { EXEC_CONVERT_SWAGGER } from "./shared/constants/SendMessageTypes"
+import { ExecConvertSwaggerMessage } from "./shared/types/SendMessage"
 
 // TIPS: background script の console.log の出力先は、「バックグラウンドページ」
 chrome.browserAction.onClicked.addListener((tab) => {
