+++ swagger-viewer/app/scripts/contentscript/app/App/fixtures/SwaggerJson_2_0.ts
