+++ swagger-viewer/app/scripts/contentscript/app/App/fixtures/SwaggerYaml_2_0.ts
