+++ swagger-viewer/app/scripts/contentscript/app/App/fixtures/SwaggerYaml_3_0.ts
