+++ swagger-viewer/app/scripts/contentscript/app/App/index.test.tsx
