+++ swagger-viewer/app/scripts/contentscript/app/App/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import SwaggerUI from "swagger-ui-react"
 import "swagger-ui-react/swagger-ui.css"
-import { MaybeSwaggerJson } from "../../types/Swagger"
+import { MaybeSwaggerJson } from "../../../shared/types/Swagger"
 
 export interface AppProps {
   /**
