+++ swagger-viewer/app/scripts/contentscript/app/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import ReactDOM from "react-dom"
-import { APP_RENDER_ID } from "../constants/App"
-import { MaybeSwaggerJson } from "../types/Swagger"
+import { APP_RENDER_ID } from "../../shared/constants/App"
+import { MaybeSwaggerJson } from "../../shared/types/Swagger"
 import { App } from "./App"
 
 export const render = (swaggerJson: MaybeSwaggerJson | string) => {
