+++ swagger-viewer/app/scripts/contentscript/util/YmlUtils.test.ts
