+++ swagger-viewer/app/scripts/contentscript/util/YmlUtils.ts
@@ -1,5 +1,5 @@
 import YAML from "yaml"
-import { MaybeSwaggerJson } from "../types/Swagger"
+import { MaybeSwaggerJson } from "../../shared/types/Swagger"
 
 /**
  * @param str yml string | json string
