+++ swagger-viewer/app/scripts/shared/__test__/Example.test.ts
