+++ swagger-viewer/app/scripts/shared/constants/App.ts
