+++ swagger-viewer/app/scripts/shared/constants/SendMessageTypes.ts
