+++ swagger-viewer/app/scripts/shared/types/SendMessage.ts
