+++ swagger-viewer/app/scripts/shared/types/Swagger.ts
