+++ swagger-viewer/app/scripts/shared/types/declare.ts
